refactor(data): 重构多模态数据处理流程
- 移除 qa_generatorV3.py，将图片识别功能整合到 qa_generatorV2.py
- 重构 ImageToTextProcessor 类，优化图片描述逻辑
- 更新数据处理配置，增加 vision_api 相关设置
- 调整数据集输出路径和格式
