+++ dataset/res_csv/sft/dataset_info.json
@@ -8,7 +8,7 @@
         }
     },
     "wechat-sft-with-history": {
-        "file_name": "sft-my-with-history.json",
+        "file_name": "sft-my.json",
         "columns": {
             "prompt": "instruction",
             "response": "output",
@@ -23,20 +23,6 @@
             "messages": "messages",
             "system": "system",
             "images": "images"
-        },
-        "tags": {
-            "role_tag": "role",
-            "content_tag": "content",
-            "user_tag": "user",
-            "assistant_tag": "assistant"
-        }
-    },
-    "wechat-img-rec-sft": {
-        "file_name": "sft-my-img-rec.json",
-        "formatting": "sharegpt",
-        "columns": {
-            "messages": "messages",
-            "system": "system"
         },
         "tags": {
             "role_tag": "role",
